refactor(gui): extract responsive layout contract tests to sibling file
primer-gui/src/responsive_layout_contract.rs carried a single inline
`#[cfg(test)] mod tests` (656 → 33 lines, −623) holding 28 frontend
contract tests that `include_str!` the static UI assets and assert the
940px mobile-drawer breakpoint, backdrop, matchMedia toggle, and ARIA
modal shape.

Move the module verbatim into the non-root sibling
`src/responsive_layout_contract/tests.rs` (mkdir). The test module is a
descendant, so `use super::MOBILE_BREAKPOINT_PX` resolves unchanged with
no visibility edits. The three `include_str!` paths gain one `../`
(`../ui/` → `../../ui/`) because the file now sits one directory deeper.

This is a default-build extraction the prior brief had marked exhausted;
the full sweep surfaced it. Pure test-only relocation, no behaviour
change. All 28 responsive_layout_contract tests preserved; primer-gui
lib suite and workspace tests unchanged.
